fix(agents): rename ieso.worldview.py to ieso_worldview.py
  The dotted filename was unimportable as a Python module — Python
  treats dots as package separators, not file-name parts — so
   in
  agents/__init__.py raised ModuleNotFoundError and the whole
  agents package failed to load.

  Rename to ieso_worldview.py and update the matching import line
  in agents/__init__.py. No behavior change beyond making the
  package importable.
